Moved assign excitation functions to Circuit level.
